Added tox target to compile kiwi into binary
With the use of nuitka it is straight forward to compile
kiwi into a binary. However, the result is not a static
binary and still a dynamic executable loading libc(and friends)
and also libpython. To prevent dynamic loading of libpython
a static version of libpython is needed. Usually this is
not provided by the distributions. All this is just a
side project to see how "interpreter free" kiwi can be