Add post_disk_build.sh script hook
The optional script hook is executed for the disk image type
oem only and runs after the image build process has been completed.
The root environment for the script is the host environment
from which kiwi was called. The main use case for this script
hook is to post process the result raw disk image file. At call
time it receives the file name of the .raw file as parameter.